Debugging

+++ src/copass_kernels.h
@@ -638,8 +638,8 @@ my_search_block_up( ArrayT array, position_t size, KeyT val, position_t* num_up 
 
   if (tid == 0) {
     printf("bid:%d tid:0 step:%ld size:%ld\n", blockIdx.x, step, size);
-    printf("arr[n-1]: %d arr[n-2] %d val %d\n", getKey(array, size-1),
-	   getKey(array, size-2), val);
+    printf("arr[n-1]: %d arr[n-2] %d val %d\n", (int)getKey(array, size-1),
+	   (int)getKey(array, size-2), (int)val);
   }
   __syncthreads();
   while ( step > 1 && ( right - left ) > 1 )
@@ -664,7 +664,7 @@ my_search_block_up( ArrayT array, position_t size, KeyT val, position_t* num_up 
       {
         shared_array[ tid ] = getKey( array, pos );
         printf("bid:%d tid:%d sa:%d pos:%ld\n", blockIdx.x, tid,
-	       shared_array[tid], pos);
+	       (int)shared_array[tid], pos);
       }
     }
     __syncthreads();
